requirements.txt
Django
djangorestframework
django-bootstrap4
django-allauth
django-cors-headers
django-dotenv
xlrd
pandas
django-protected-media
Pillow
unidecode
